abstract base class
- made Application & NodeBase abstract
- moved A & B specific code to child classes
- apply code format
