+++ src/SparkplugNet/VersionA/SparkplugNode.cs
@@ -9,6 +9,8 @@
 
 namespace SparkplugNet.VersionA;
 
+using SparkplugNet.Core;
+
 /// <inheritdoc cref="SparkplugNodeBase{T}"/>
 public class SparkplugNode : SparkplugNodeBase<VersionAData.KuraMetric>
 {
@@ -20,5 +22,103 @@ public class SparkplugNode : SparkplugNodeBase<VersionAData.KuraMetric>
     /// <param name="logger">The logger.</param>
     public SparkplugNode(List<VersionAData.KuraMetric> knownMetrics, ILogger? logger = null) : base(knownMetrics, logger)
     {
+    }
+
+    /// <summary>
+    /// Publishes version A metrics for a node.
+    /// </summary>
+    /// <param name="metrics">The metrics.</param>
+    /// <exception cref="ArgumentNullException">The options are null.</exception>
+    /// <exception cref="Exception">An invalid metric type was specified.</exception>
+    /// <returns>A <see cref="MqttClientPublishResult"/>.</returns>
+    protected override async Task<MqttClientPublishResult> PublishMessage(List<VersionAData.KuraMetric> metrics)
+    {
+        if (this.options is null)
+        {
+            throw new ArgumentNullException(nameof(this.options), "The options aren't set properly.");
+        }
+
+        if (this.KnownMetrics is not List<VersionAData.KuraMetric> knownMetrics)
+        {
+            throw new Exception("Invalid metric type specified for version A metric.");
+        }
+
+        // Remove all not known metrics.
+        metrics.RemoveAll(m => knownMetrics.FirstOrDefault(m2 => m2.Name == m.Name) == default);
+
+        // Remove the session number metric if a user might have added it.
+        metrics.RemoveAll(m => m.Name == Constants.SessionNumberMetricName);
+
+        // Get the data message.
+        var dataMessage = this.MessageGenerator.GetSparkPlugNodeDataMessage(
+            this.NameSpace,
+            this.options.GroupIdentifier,
+            this.options.EdgeNodeIdentifier,
+            metrics,
+            this.LastSequenceNumber,
+            this.LastSessionNumber,
+            DateTimeOffset.Now);
+
+        // Debug output.
+        this.Logger?.Debug("NDATA Message: {@DataMessage}", dataMessage);
+
+        // Increment the sequence number.
+        this.IncrementLastSequenceNumber();
+
+        // Publish the message.
+        return await this.Client.PublishAsync(dataMessage);
+    }
+
+    /// <summary>
+    /// Called when [message received].
+    /// </summary>
+    /// <param name="topic">The topic.</param>
+    /// <param name="payload">The payload.</param>
+    /// <returns>
+    /// A <see cref="T:System.Threading.Tasks.Task" /> representing any asynchronous operation.
+    /// </returns>
+    /// <exception cref="System.InvalidCastException">The metric cast didn't work properly.</exception>
+    protected override Task OnMessageReceived(string topic, byte[] payload)
+    {
+        var payloadVersionA = PayloadHelper.Deserialize<VersionAProtoBuf.ProtoBufPayload>(payload);
+
+        if (payloadVersionA is not null)
+        {
+            var convertedPayload = PayloadConverter.ConvertVersionAPayload(payloadVersionA);
+
+            if (topic.Contains(SparkplugMessageType.DeviceCommand.GetDescription()))
+            {
+                if (convertedPayload is not VersionAData.Payload convertedPayloadVersionA)
+                {
+                    throw new InvalidCastException("The metric cast didn't work properly.");
+                }
+
+                foreach (var metric in convertedPayloadVersionA.Metrics)
+                {
+                    if (metric is VersionAData.KuraMetric convertedMetric)
+                    {
+                        this.DeviceCommandReceived?.Invoke(convertedMetric);
+                    }
+                }
+            }
+
+            if (topic.Contains(SparkplugMessageType.NodeCommand.GetDescription()))
+            {
+                if (convertedPayload is not VersionAData.Payload convertedPayloadVersionA)
+                {
+                    throw new InvalidCastException("The metric cast didn't work properly.");
+                }
+
+                foreach (var metric in convertedPayloadVersionA.Metrics)
+                {
+                    if (metric is VersionAData.KuraMetric convertedMetric)
+                    {
+                        this.NodeCommandReceived?.Invoke(convertedMetric);
+                    }
+                }
+            }
+        }
+
+        return Task.CompletedTask;
     }
 }
